18 support tags for series

+++ Models/ManOfTheSeries.cs
@@ -11,7 +11,7 @@ namespace Com.Dotnet.Cric.Models
         public long Id { get; set; }
         [Required]
         [ForeignKey("Series")]
-        public long SeriesId { get; set; }
+        public int SeriesId { get; set; }
 
         public Series Series { get; set; }
 
@@ -27,7 +27,7 @@ namespace Com.Dotnet.Cric.Models
             // Parameterless constructor required by Entity Framework Core
         }
 
-        public ManOfTheSeries(long seriesId, long playerId)
+        public ManOfTheSeries(int seriesId, long playerId)
         {
             SeriesId = seriesId;
             PlayerId = playerId;
